Extract travis into own module
